Hardcode models with partial sinks for string formatting functions in the backward analysis
Summary:
The previous diff hardcodes the models in the forward analysis. Now it is time
to hardcode in the backward analysis.

This diff also increases the trace lengths in the backward taint when calling
string format functions, to reflect the fact that such invocation increases the
trace lengths. It is a bug that we did not increase the trace length before.

Although we manage to reintroduce the model for
`DatabaseSchemaEditor.string_operations`, the model is still not precise:
1. We should infer that `self.attribute_query` has a sink taint, but we instead
infer `self` has a sink taint. This is a pre-existing problem, as demonstrated
in `stradd_triggered_user_controlled`.
2. The triggered sink should have a call info of `Origin` instead of
`Declaration`, such that the UI can highlight where the sink taint is
introduced.

These problems will be addressed in the follow-ups.

Reviewed By: alexkassil

Differential Revision: D43674767

fbshipit-source-id: 048429bda98a1a5f0027f7f8424ebe0f2fe6ec10