Remove warning for lack of targets
Summary:
With the lazy server not requiring targets, it doesn't make sense to show this warning anymore.

If we do want to show it, we should first check that we are in non-lazy mode. We don't have access to that at this point so I will leave it out for now.

Reviewed By: grievejia

Differential Revision: D43710413

fbshipit-source-id: 5067bdda5c4449c5c7c5bfa67b5d39dd3ab14ef6