Allow the typeshed downloader to use a directory for output
Summary:
I'm leaning toward saving our patched typeshed stubs as a vendored directory
rather than a committed zipfile, for two reasons:
- on trunk, it would make reading the stubs easier and code search possible
  (and hence debugging stub issues much easier for users).
- when debugging errors after a typeshed change (both new errors popping up
  in our sanity checks against internal repos and pysa model verification
  problems) it would be *vastly* easier to have trivial access to a diff
  of the old vs new typeshed - this is something we don't get from zipfiles.
  Yes, it's possible for me to diff the original typeshed repos on two
  commits or to build tools on top of the zipballs, but that's much harder
  than if I could get a diff view directly on my commit.
  - This second point is really important to me; we've been delaying typeshed
    upgrades pretty frequently because they are so painful. Anything that makes
    them easier (and a diff view sure would!) will be a big help and make it
    more likely that we stay up to date :)

This commit starts moving in that direction by allowing the stubs to be written
out to a directory instead of a zip file. For now we still use a zip file by
default.

Even if we don't wind up moving forward with this, I think having a directory
output option is nice for debugging - it saves us from having to look up the
linux command for extracting a zipball in order to look at the patch results.

Reviewed By: connernilsen

Differential Revision: D43708430

fbshipit-source-id: 08fbc78ceeea6224949c1af2ab02f0338c7b6e32