Remove Dockerfile
Summary:
Apparently this dockerfile was created to deal with build issues
when M1 macos machines first came out. The problem is that it's
not checked by CI and is not maintained - for example the ocaml version
is out of date.

Rather than updating it to reflect typeshed changes I think we
want to just remove it. We could revive it in the future; if we do
I think we should make sure there's a github CI action to stress-test
it so that we can be sure it keeps working.

Reviewed By: grievejia

Differential Revision: D43774847

fbshipit-source-id: bea180bc21220e36ebd467d4490dee891bad1c3f