Rename vendored typeshed directory
Summary:
It is not clear why we historically named the `typeshed`
directory to `typeshed-master` in the zipball, only to rename
it to `typeshed` in all downstream code.

Life will be easier if we don't do this - among other things,
then our vendored typeshed will actually be a valid typeshed
that one could point to from a pyre configuration, which will
be much nicer for integration tests and for manually running
against distillery when working on patches!

Note that at this stage downstream code still uses the zipball
and typeshed-master, I'm just renaming the directory version
to make flipping over more convenient. The code is a little
clunky because of the need to still use `typeshed-master` in
our zipball, but I'll clean it up after I remove the zipball.

Reviewed By: grievejia

Differential Revision: D43774297

fbshipit-source-id: 38bceed7f49290c6476171bb058434fb9869603b