Adds support for checking python version in models
Adds support for checking python versions in pysa models by extending
the model parser to parse if-conditions as well.

Previously all if-conditions in the modesl file were discarded with a
model verification error. Extend the AST logic to check for the presence
of if conditions and honour them. Currently, we can check against the
python version in the system provided to the taintAnalysis Command.

All comparision operators except in and not in are supported.
If the right side expression is not a tuple, the condition is ignored and
a model verification error is emitted. Same as if the left side of the
condition is not sys.version or if the condition doesn't have two sides
or uses a supported operator.

Updated the docs (website) to specify the same.

Updated existing tests to use default python version and added unit
tests to modelTest.ml

Signed-off-by: Abishek V Ashok <abishekvashok@gmail.com>