Make sure patch directory exists in download_typeshed.py
Summary:
On multiple occasions while doing the latest typeshed bump, I would put up a
diff with an incorrect typeshed because I messed up the patch directory
argument - in some cases because I was using a relative path from an earlier
exeuction when running from a different directory, and in other cases becasue
of typos.

This is a really easy mistake to catch, let's just make sure we complain if
the patch directory doesn't exist!

Reviewed By: grievejia

Differential Revision: D43879016

fbshipit-source-id: 01ac298ace52250e6282765e59769e570152c9a6