UninitializedLocalCheck should not skip attribute bases
Summary:
There is a known and hard-to-fix bug in UninitializedLocalCheck caused by
qualification, where *direct* attribute accesses on uninitialized locals are
difficult to check because we don't clearly distinguish between accessing
attributes on a name defined in the current scope versus accessing some global
name that is part of a package.

But, several of the bug reports we've had related to this are actually caused
by failing to look at the attribute base at *all*, which causes us to miss
easy-to-catch cases such as using an uninitialized variable inside a chained
method call.

This commit fixes that problem, by setting a flag that disables recording
`Name.Identifier` reads on *direct* ancestors of a `Name.Attribute`, while
still checking inner expressions.

Reviewed By: grievejia

Differential Revision: D43890452

fbshipit-source-id: 9c81a9cde8225c592b92b0b10a6e99941d8a800d