Add a new example to uninitialized local check
Summary:
This example doesn't look like an attribute access, but because we desugar
binary operators it is one under the hood. This wasn't instantly obvious
to me (and it shouldn't be - if we got rid of eager desugaring in the AST
this would be fixable without qualification changes!), so I think having
an actual test case is helpful.

Reviewed By: grievejia

Differential Revision: D43890860

fbshipit-source-id: 2fa5eaca7929335d50310ea51aebfc97313f1d5a