Rearrange `from_comprehension` code
Summary:
I was trying to skim the uninitialized local check code, and found
it a little difficult to understand `from_comprehension` at first,
just because I got lost on the arguments.

Using a named argument and re-ordering the arguments so that `collected`
goes first (that's the more common ordering in Pyre for a resursive
fold operation like this) helps a bit.

Reviewed By: grievejia

Differential Revision: D43890453

fbshipit-source-id: 6f3393a7fc4a52de07141c112f0a728888c0e311