Add test case for uninitialized local uses in attribute base
Summary:
This commit adds a pair of tests illustrating that Pyre fails to catch
uninitialized local reads happening in the callee position of method calls;
the underlying reason is that we skip rescursion on the base part of a
`Name.Attribute` type expression.

This limitation is already recorded in T94414920 and has tests that
direct attribute access is not caught, but we want additional tests because
many uses of uninitialized locals that *aren't* attribute accesses are
in principle catchable without a qualification refactor.

Reviewed By: grievejia

Differential Revision: D43890455

fbshipit-source-id: 26cd087fa8887fdd4abeecf078510c2b49659d51