Make numbers.Number non-abstract
Summary:
The latest typeshed has gotten rid of the explicit `__hash__`
function, among others, from primitive classes - these aren't
really needed because all primitives inherit from `object`.

But Pyre hacks the type hierarchy (inside of hardcoded hooks
in `classHierarchyEnvironment.ml`) to force these primitives to
inherit from `numbers.Number`; this inheritance is truthy in
the sense that they don't *actually* inherit but we do want
the typing relationships to show that, and `isinstance` behaves
as if they did.

This causes a problem though because `numbers.Number` declares
`__hash__` to be abstract, which in turn causes primitive types
to be abstract, because in our hacked inheritance graph
`Number` comes before `object` in the method resolution order.

I looked at fixing this inside of Pyre but ran into some roadblocks;
by far the easiest fix is to just make `Number` non-abstract.

Reviewed By: jasontatton

Differential Revision: D44034073

fbshipit-source-id: 82c53737e455d21be5eeb7303ee15e21b3aa8f87