Update test stubs
Summary:
In upcoming diffs, I want to add a test of Pyre's actual behavior
against primitive numeric types, given the stubs in typeshed.

The relevant behavior is entirely dependent on details of the stubs,
so I need to first update the test stubs to be in sync with
typeshed.

Note that no ocaml code (other than tests) changed at all, so we can
be pretty confident that all changes are just due to the stubs changing.

Reviewed By: grievejia

Differential Revision: D44030639

fbshipit-source-id: 8ff8ddd5bf251ae255802f5e343fb7961e2fd7ac