Add a test where we inherit from a primitive numeric type
Summary:
This kind of inheritance happens in some generated thrift
enum code, and as a result, our behavior here is important -
there can potentially be hundreds of instantiations of these
enums, so marking each instantiation (incorrectly) as
an illegal abstract class instantiation is a bug.

I don't want to get into why exactly we have this bug in
this diff - the following diff (where I fix the problem)
is a better place for that discussion. This commit simply
documents the bug in a unit test.

Reviewed By: jasontatton

Differential Revision: D44034074

fbshipit-source-id: 0e5eef4944fcf32d5f4873f700510a29e63bc8bb