Cache File Opening
Summary:
The pyre linting protocol has the following synchronous calls for every lint rule:
- start pyre server
- open file
- query for superclasses

Only the 3rd call is necessary on every visit if we are visiting within the same file. Since there can be many visits in the same file, we should cache the files we open to not block on repeat open file requests.

Reviewed By: grievejia

Differential Revision: D44066182

fbshipit-source-id: c702cbb9efb141f4caa7ddcb554f59a6233a1089