Treat a nested string expression as TITO when there is no stringify targets
Summary:
Such that the taint on the nested string expression can be preserved.
Otherwise, the taint will be lost.

Meanwhile, we intentionally do not analyze f_string_targets on the nested
string expression, because such targets are artificial and do not have actual
meanings -- they are introduced only to distinguish f-strings at different
locations in the same function.

Reviewed By: arthaud

Differential Revision: D44096180

fbshipit-source-id: 3e8f074f4861219af81ffc0588d306e19dd669b2