Improve the API for inlining decorators
Summary:
The inlining decorator module exposes a lot of internal functions.
This diff improves its API by clearly exposing at the top the function one could use to control inlining of decorators.

In the future, we will refactor it further to add the ability to completely remove some decorators that lead to false negatives in Pysa.

Reviewed By: tianhan0

Differential Revision: D44101610

fbshipit-source-id: cb4abff3dafdf63ac862eb55bd9c0d6a96b38049