Pin typeguard to earlier version in client tests (#726)
Summary:
Pyre open source builds are failing, and I've tracked down the issue to a recent major API change with typeguard, one of testslide's dependencies.

As a temporary fix, I've pinned the version of typeguard we use to the version prior to the major update, which seems to have fixed the issue. I'll also be notifying the typeguard team so they can hopefully get a fix out for their OSS customers.

Pull Request resolved: https://github.com/facebook/pyre-check/pull/726

Reviewed By: stroxler

Differential Revision: D44111937

fbshipit-source-id: 2833de95eeac1c940bedb36c0d78235c7cd8384e