Emit readonly error if an incompatible-parameter error is related to readonly.
Summary:
Problem: When deciding whether to emit a readonly error, we check if the actual type is `ReadOnly[X]` and the expected type is non-readonly. This missed readonly mismatches deeper within a type (e.g., `Iterable[ReadOnly[X]] vs Iterable[X]`) and emitted a generic "incompatible parameter error".

That makes life harder for the readonly automation script (for example, we don't always get the fully-qualified name with the non-readonly error). and makes it harder to track readonly-specific fixmes.

Solution: Check if the error is related to readonly by stripping `ReadOnly` and seeing if the types are compatible.

This means that we will get the fully-qualified name in the automation script.

Design alternative: While we do want the above fix to solve the problem of specific error codes, another way to get the qualified name for the script is to make the `long_description` field of the JSON error output always have the qualified name. Problem: The types have been dequalified (in postprocessing) long before we reach [`AnalysisError.instantiate`](https://www.internalfb.com/code/fbsource/[e2e2236a511e]/fbcode/tools/pyre/source/server/requestHandler.ml?lines=16). So, that's not something we can do right now, unless we set up another query or flag to emit fully-qualified names. We can look into that later on if we encounter the problem a lot.

Reviewed By: grievejia

Differential Revision: D44114145

fbshipit-source-id: 6b2b11b29db87562c78aeb8592aa4064cd41a18e