Introduce a new mode `@IgnoreDecorator`
Summary:
This diff introduces a model mode `IgnoreDecorator` which can be used to completely ignore a specific decorator during the analysis.
This is useful when a decorator adds logic that makes the analysis less precise and introduces false negatives.

For instance:
```
def my_unannotated_decorator(f):
  return f

my_unannotated_decorator
def my_function(command):
  os.system(command)
```
With the unannotated decorator, we will most likely not find flows to `os.system`.

This diff only introduces the parsing of the new mode. The implementation will be done in follow-ups.

Reviewed By: alexkassil

Differential Revision: D44131510

fbshipit-source-id: 20cb157bcce668ae333c6a82619f46480495b955