Use balanced fold for joining override models
Summary:
When joining many overrides, it doesn't matter the order the join happens.

So with n being length of list of overrides, and m being the cost of one `Model.join` call, the regular fold solution is `O(n*m)`, while using the `fold_balanced` solution results in `O(log(n)*m)` by joining models pairwise, and then joining those pairs, then those groups of initially 4 models, etc.

`Algorithm.fold_balanced`'s type signature is `'a list -> f:('a -> 'a -> 'a) -> init:'a -> 'a` as opposed to `List.fold`'s `'a list -> init:'accum -> f:('accum -> 'a -> 'accum) -> 'accum`. So we need to convert override targets to models before we can swap to the efficient fold.

Reviewed By: arthaud

Differential Revision: D44154858

fbshipit-source-id: f3489446ffaa7e2ea12d4b0bafd3e949f269ff91