Improve LibCST module parsing logic
Summary:
A raw LibCST module with no location metadata is not useful
to Pyre in the context of statistics collection. So let's
just push that down, which makes the top-level code easier
to follow.

While we're at it, let's rename the functions - we're not
parsing a path but rather parsing the code *at* a path,
so `module_from_path` is a nicer description than
`parse_path_to_module`.

This required some minor updates in the coverage interface
to reflect that we always want to use a wrapper (before, it
was asking for a raw module and constructing the wrapper
itself).

Reviewed By: grievejia

Differential Revision: D44228464

fbshipit-source-id: 42904b93023327eed0df247b5b35e7a0e07dae5e