Move module_from_path to libcst_collectors.py
Summary:
We always need to parse a path to a MetadataWrapper (a
libcst module + position information) before we can run a
libcst collector on it.

So these functions belong in that library module, not in
`statistics.py` - the fact that they were in `statistics.py`
was a source of coupling between commands, which is one of the
things making our code hard to navigate.

Reviewed By: grievejia

Differential Revision: D44228463

fbshipit-source-id: 8be5dfbbd0cc20bec635a7791cd6402293047069