Add logging verbosity flag to Pyre client
Summary:
The ability to set log levels was requested in a Q&A question for a script another team is writing. This team would like the ability to filter logging to only be error levels or above.

This diff adds support for the default log levels provided by python3's builtin `logging` module.

Reviewed By: stroxler

Differential Revision: D44032697

fbshipit-source-id: 2297b171545755e26ebf05848a97a77262ff3633