Simplify `find_roots` a bit
Summary:
In the previous diff, I made `pyre statistics` throw if the
user tries to collect statistics on a directory that isn't part
of the project.

I made the diff minimal by replacing some logging with a raise in-place,
but we can make the function easier to skim by moving this condition
up, so that the last line in the loop over
`explicitly_specified_directories` is adding to `absolute_paths`; this

Similarly, I think it's easier to see control flow if we move the simpler
branch, where there are no explicitly specified directories, up to the
top.

Finally, the local and global roots are always absolute (they are coming from
configuration, which only deals in absolute roots) so there's no need for
and `absolute_root` variable (and even if they were, just calling the Path
constructor on an existing Path wouldn't make it absolute!)

Reviewed By: grievejia

Differential Revision: D44232434

fbshipit-source-id: c929fb77ae6ea451fd0a50499ecd1aee9514093e