Inline the `.py` suffix check
Summary:
I'm all for layering functions in cases where either they
are conceptually very different *or* the control flow is nontrivial.

But in this case the `_should_ignore` function is already a predicate
that looks at the path's "shape" as a string, and inlining the ".py"
check on the suffix does not meaningfully increase complexity - it's
all one big `or` chain.

So there's not much benefit to the extra layer, let's just strip it
out.

Reviewed By: kinto0

Differential Revision: D44242541

fbshipit-source-id: 5df34ddf6901c0ab4fb1f40560cd91420e049d25