Rename some functions and attributes and improve types
Summary:
The `pyre statistics` command does not actually take a list
of directories as input, even though the interface suggests
it does.

Instead, it takes an *optionl* list of *paths* (which may
be directories or regular files).

This diff tweaks our code to reflect this, by:
- converting the `directories` click argument to be an optional
  list of paths instead of a required list of strings
- renaming the variables to reflect that they can be file
  or directory paths
- renaming the find_roots function to reflect that it is finding
  top-level paths, which aren't *necessarily* roots (or even
  directories, they might just be plain files)

None of the semantics actually change, the names and types just
now communicate what we're doing more clearly.

Reviewed By: kinto0

Differential Revision: D44235036

fbshipit-source-id: 931e45eb778878f82ba48fbf0097680bcfed116a