Kill off LSP support for function-level coverage
Summary:
Function-level type coverage has mostly been a dud in LSP.

Two factors are at play here:
- It's not a great idea to run LibCST visitors inside the LSP, because
  they will block the current thread, so trying to get function-level
  coverage causes nontrivial perf issues
- All it does is tell you which functions have annotations. This is trivial
  to see by eye, it's not actually helpful to highlight in blue.

In addition, expression level coverage is potentially *very* useful, and
it's very complicated to configure switching between them in a sane way.

So it seems strictly better for us to always use the more useful and
more performant option, which is expression-level coverage, rather than
keeping around a not useful and mostly-dead feature in a fast moving
and complex codebase.

Reviewed By: kinto0

Differential Revision: D44242545

fbshipit-source-id: cde5484a9cf34c93ec06b8fc0aa91b475041888a