Explicitly annotate list that might be empty
Summary:
I was trying to double-check that we don't need some legacy
coverage-related types from the `coverage_data` module, so I
was hovering on some code in `daemon_querier` when I realized
that this variable is any-typed!

The problem is that the first assignment is assigned an empty
list, so without function-level constraint solving Pyre just
has to give up.

Adding an explicit annotation removes the expression-level
coverage gap (and makes hover work for me :) )

Differential Revision: D44271390

fbshipit-source-id: 55c75a9b4a0a67c2f789a6abefca3868cd7a65c0