Add AnnotationClassExtends clause to AnnotationConstraint
Summary:
This adds a new `extends` clause to `AnnotationConstraint`. This is motivated by wanting a `return_annotation.extends()` constraint - for example, we may want to taint the return value of all methods returning something that extends a `FooResponse` class as a sink.

Note that this should only work on classes. For cases like `Optional[T]` and `ReadOnly[T]` these will resolve to just `T`.

Reviewed By: arthaud

Differential Revision: D44146476

fbshipit-source-id: 635a77d27b7386612e0819df6e105e442fb5c8bf