Isolate legacy coverage tooling in `coverage.py`
Summary:
At this point I'm about ready to begin work on a new coverage command that
will expose similar data to `pyre statistics` and `pyre coverage`, but
in a more usable format (in particular, I want the key information about what
the problems are that comes from `pyre statistics`, but with line numbers that
are only available in `pyre coverage` today).

We don't want to remove the `pyre coverage` command yet because at this point
it's the only tool users have that even tries to report line numbers. But
I don't want those tools - which we can remove as soon as we have better data -
cluttering the `coverage_data.py` library module. Instead, I want to encapsulate
all that code in `coverage.py`.

That's what this diff does, which shrinks both the production and test codebases
for `coverage_data`.

Reviewed By: grievejia

Differential Revision: D44304811

fbshipit-source-id: 42d83dae623449048475befbe54541cf86be8662