Move expression-level coverage logic down
Summary:
For reasons that are unclear to me, we were doing a decent chunk of the business logic
for expression-level coverage inside of the top-level entrypoint to the internal client.

This doesn't make sense to me, logic like converting raw user-provided paths into
the actual paths we'll send to the backend belongs at a lower level.

Reviewed By: grievejia

Differential Revision: D44314880

fbshipit-source-id: 9861b0bf112088388bea7007241aa895a4bef392