Make changes to expression-level coverage path handling
Summary:
I wasn't sure exactly how to title this diff because I dont' think it's
fair to say I'm "simplifying" or even necessarily "improving" the path
handling, but what I'm doing is:

- Making the path handling much more explicit, and encoding the distinction
  between argument files vs normal paths in the type system.
- Using code from `coverage_data` to resolve paths instead of from `coverage.py`;
  the former is intended to be used as a library whereas the latter is not.

The new code is a little ugly, it feels over-engineered to have a whole dataclass
just to convert one list to another list, but having the intermediate state described by a type makes it much more explicit that handling of argument paths vs
normal paths is completely different.

I think ideally this logic should be pushed out into `coverage_data.py` so that
other coverage commands can potentially reuse the arguments file logic, but
in this diff my main motivation is just to get rid of the coupling to `coverage.py`.
I'll try to clean it up in the process.

Reviewed By: grievejia

Differential Revision: D44316771

fbshipit-source-id: f204ad4987661c3e581aa8f51ec06f39199cc422