Tighten up tests of CoveragePaths
Summary:
The CoveragePaths tests were in a weird state because they used
a fake project that had no actual code, but the code they are testing
actually reads from the filesystem.

As a result, passing empty paths resulted in nothing in unit tests,
which was misleading about the actual behavior - when empty paths
are passed, we will get coverage for all files in the project!

In addition, there was a bug in my initial implementation of the previous
diff which I didn't discover until end-to-end testing: the initial
code would get coverage for all paths in the project even if an arguments
file was passed, and the unit tests couldn't catch this because there were
no code files in the project!

A fix for both of these problems is to touch a few python files in the test.

Reviewed By: grievejia

Differential Revision: D44317756

fbshipit-source-id: 1b65a8a91f7f88b5c978bd855fbdb3b4d0072d0f