Integration: Improve return type from parse_module
Summary:
parse_module will now return a:
```
module ParserError : sig
  type t =
    | Recoverable of {
        recovered_ast: Ast.Statement.t list;
        errors: SyntaxError.t list;
      }
    | Unrecoverable of string
end
```

This is a much better return type as it permits the consumer flexibility in how the returned value is handled (recoverable vs not etc).

Instead of a string. Unit tests have been updated to reflect this change in return type.

Reviewed By: stroxler

Differential Revision: D44296775

fbshipit-source-id: e2c9cb6465a9ef863bc1befbcfa1403a8dd5a6f9