Refactor global leak check to pass reachable globals up instead of mutable expressions down
Summary:
**Background**: In early development of global leak check, a quick decision was made to determine if data is being leaked to a global by passing a boolean flag down if a mutable expression (known mutation method or assignment statement) was encountered. We would then error if/when a global was reached. This was an easy to implement solution that covered the simple cases we were encountering at the time.

In further development, a few hacky solutions had to be implemented to keep the structure the same and cover edge cases. Now, it's time to go through and fix up these solutions to provide a cleaner and more extensible solution.

Instead of passing a boolean flag down, we will now pass reachable globals upward (along with other metadata that may be useful), and filter them out as we go. We will now error if a reachable global makes it to a mutable expression.

Reviewed By: pradeep90

Differential Revision: D44337351

fbshipit-source-id: 5345cee80a108f117482c4668cfa060e28724444