Rename StrictCountCollector -> ModuleModeCollector, ModuleStrictData -> ModuleModeInfo
Summary:
Names kind of matter for keeping track of things, and what we're collecting
when we look at mode information is the module mode - it's not best named
"strict data" because "strict" is just one of the three modes.

So, rename both the dataclass tracking mode information and the collector
(which also wasn't really *counting* anything, at least once we removed the
zombie count attributes that weren't being used in the first place).

Reviewed By: grievejia

Differential Revision: D44399005

fbshipit-source-id: e9c62c1dfcd0eb99353eb757559fec29a7187efb