Add tests of CoverageCollector
Summary:
Up until now, all of the tests were on CoverageCountCollector, but
I don't want to keep using that interface - it does not include enough
information for building good UX or good integration with tools that need
names, locations, etc.

That means we need tests that actually use the FuncitonAnnotationInfo
output of the raw collector, which is what I actually want to build on.

Reviewed By: grievejia

Differential Revision: D44397087

fbshipit-source-id: 18fdda702343eabe2e56039dba894a4108f7ae48