Move legacy suppression logic to statistics.py
Summary:
justmovingthingsaround

For reasons discussed in the previous diff, the existing suppression-finding
code used by `pyre statistics` is:
- not very helfpul for producing the most useful kind of tablular data that
  will meet various needs (such as tracking suppressions over time or displaying
  them nicely in a UI)
- not worth the effort of rewriting in terms of tabular data

The previous diff re-implemented suppression finding. So at this point,
the legacy code is not very interesting and is exclusively used by `statistics.py`.

We still need it until we've completely eliminated uses of the old data, but
having zombie code like this sitting in a core module like coverage_data.py makes
it harder to see what code matters or to move quickly on changes we want to make.

Moving it into `statistics.py` solves these problems - the code is unchanged, the
full test harness is still there, but it's now out of the critical path in
`coverage_data.py` where we are actively making improvements.

Reviewed By: grievejia

Differential Revision: D44398725

fbshipit-source-id: 34b50a821bb03b3df6d8b4a2f4fe34712cf79176