Move AnnotationCountCollector to `statistics.py`
Summary:
The rationale here is the same as for the previous diff moving legacy
suppression-count logic: it isn't actually all that useful to produce counts
inside of Pyre itself, instead we want nice tabular data with all the
information we would want (code ranges, names of variables, etc) for reporting.
We can easily aggregate that up in data pipelines later, we don't particularly
need aggregated data coming out of the Pyre CLI.

As a result, the AnnotationCountCollector, which we do still need for as long
as we have use cases of `pyre statistics`, is not going to be useful going
forward and it's better to move it out of `coverage_data` and into `statistics`
so that it is out of the way as we work on implementing an improved coverage-data
collection command.

Reviewed By: grievejia

Differential Revision: D44398861

fbshipit-source-id: e3aec69534728893799c96a20a76ad43d1a2a452