Replace AnnotationInfo for parameters, returns
Summary:
I'm working on making coverage data actually useful, which means having enough metadata
to answer the questions we would want to, such as: what are the names of untyped parameters
or functions with untyped returns?

The context needed to answer these questions varies slightly depending on what kind of
thing we are annotating - for example we want both the parameter and function name for
parameters, but just the function name for returns. As a result, having similar but
distinct dataclasses helps us present a well-typed interface.

This commit flips just parameters and returns to use new logic.

**NOTE ABOUT ATTRIBUTES AND GLOBALS**

I'm planning to skip attributes and globals because, after reviewing the both the
quality of the signals (very poor - for example because the frontend doesn't do
scope analysis we can't distinguish between an unannotated value vs a value with
an explicit annotation and then later an assignment) and use cases (most of the
users of coverage data completely ignore attributes and globals) I strongly
believe we should leave them out of the next iteration of coverage entirely.

I'll leave the existing logic as it stands for now, but start stripping attributeds
and globals from downstream code; once all references to it are gone I'll delete the
code entirely.

If we care about attribute and global signals, I believe that we must move coverage
into the backend. I will not spend my time trying to maintain this logic in the frontend,
the signal quality is too bad and my time is too valuable.

Reviewed By: grievejia

Differential Revision: D44382185

fbshipit-source-id: dcd93a976c83030a7889a393ebb0322f4592adf1