Remove unnecessary class
Summary:
I'm not really sure what the StatisticsCollector class
was supposed to be for. I think it probably was there to signal
intent, that children should be used by `statistics.py` or something,
but the mixin doesn't have an interface and I don't really think
it's needed - all these classes are defined in `coverage_data` already.

Let's just get rid of it, it's better to maintain less code rather than
more code.

Reviewed By: grievejia

Differential Revision: D44392734

fbshipit-source-id: 3736e70b6b7ee5aae54be98e3476316448033a2c