Switch to using a class name stack instead of a count
Summary:
In the future, I want to be able to preserve human-readable context on
where annotation gaps are.

To do this, we shouldn't just count the class defines we enter, we should
track them. That will enable a best-effort attempt to get properly
qualified names for methods and attributes (best-effort is fine and I'm
not going to worry too much about edge cases like nested functions - today
we have nothing and it's really bad for all use cases, so just getting
the typical case right will be a huge improvement).

The current diff just replaces a count with a stack but doesn't use it yet;
I'll make use of it later when I create dataclasses representing each kind
of annotation in a later commit.

Reviewed By: grievejia

Differential Revision: D44382184

fbshipit-source-id: f63468a32d9b9af30da00f9e11beb5a63cb1bfb4