Simplify parameter handling a bit
Summary:
I'm trying to rework Pyre's coverage data so that we can sensibly
get finer-grained information about coverage gaps: where are the gaps
in terms of line numbers, what kinds are they (parameters, returns,
globals, etc), what are the names of the nodes lacking them.

All of this will help us build finer-grained datasets, and more importantly
provide useful information to users (who currently don't really have
a human-readable way to see where coverage gaps come from).

To do this, I need the data to actually make good sense, which means we
need to pick a place for handling `self` / `cls` and stick with it. Today,
we double-count it: we treat it as an annotated parameter, *and* we exclude
it from various other ways of counting things.

This creates confusion because there's no central place with responsibility.
My decision here is to consider that parameter to be unannotated unless it
has an actual annotation (note that it may be typed, but it is in fact not
"annotated").

In the process I also simplify the logic a bit - let's not double-iterate
over parameters in the visit_FunctionDef method, this is just confusing. Instead
let's output a Sequence and just pass it along. Similarly, don't pass the list
of `libcst.Param` values to `FiunctionAnnotationKind.from_function_data`, just
to count them - this is unnecessary and makes the tests way more confusing. If
all we need is the count, let's just pass the count!

Reviewed By: grievejia

Differential Revision: D44377639

fbshipit-source-id: 551cc512c725719dfd755066b5454c43d0c988f9