Improve FunctionAnnotationKind unit tests
Summary:
There's a new unit test here, covering an edge case where self is annotated but
some other parameter is not. See comments on D44377639 - the lack of this
example in the test suite let a bug slip by me on my first attempt at
simplifying the logic.

In addition, I removed the `three_parameters` variable, which was making
these tests kind of silly.

Basically, the current interface takes in both a list of parameters *and*
a count, which means it consumes duplicate information and is tightly coupled
to other parts of the code (for example the snippet that says an untyped
self parameters is annotated, which is questionable).

Then the tests compound this problem by verifying counts that are inconsistent
with the actual parameters - in other words, the tests are verifying the
code in scenarios that could never actually happen unless the rest of the
codebase were buggy!

Given that I want to clean up the implementation to not take duplicate information
in the next diff, I have to first clean up these tests to pass information in
a consistent way.

Reviewed By: grievejia

Differential Revision: D44439208

fbshipit-source-id: 81e819ba2d5870b5a90e24ff9abc4a776376b425