Add SuppressionCollector
Summary:
For several reasons, I believe the SuppressionCountCollector code
needs to be forked into `statistics.py` temporarily, and eventually
abandoned.

The current interface is not compatible with what we need:
- We want fine-grained data going into hive that can be trivially
  aggregated up in data pipelines. That data should be flat (i.e.
  a list of suppressions) because that's what data pipelines like -
  nice flat data that can be aggregatted and transformed easily
  in SQL. But the current data is deeply nested, pre-aggregated
  into a slightly complicated data structure.
- We want to at least have the *option* of using quality analyzer
  or some other off-the-shelf tool to visualize suppressions. Most of
  these tools require code ranges with columns, not just line numbers,
  but the legacy suppression collector is purely line-number based.

For these reasons, I've implemented a new TypeErrorSuppression type
representing a suppression in a flat way (so we can just have a list of
them per module) and a new SuppressionCountCollector that gets them.

As for the old SuppressionCountCollector, I was initially thinking I would
re-implement it in terms of the new one, but actually producing the
weird format it wants is not that convenient. It does not make sense to
spend valuable developer time building out that transformation when:
(a) tested code already exists that produces the format `pyre statistics` wants, and
(b) I'm pretty sure in the medium-term we just want to replace `pyre statistics`,
    so there's essentially no long-term impact to maintaining that code.

So, I'm leaving the legacy code untouched in this diff; next diff I'll move
it all to `statistics.py` so that it isn't adding line noise to the `coverage_data.py`
module and hopefully within a few months we can just delete it.

Reviewed By: grievejia

Differential Revision: D44398564

fbshipit-source-id: 2635b8b802268ab599bc4f1b965c05e46ff1d01b