Extract position metadata handling to a mixin
Summary:
Similar to the last diff, I'm addressing the fact that I don't like how
libcst visitors wind up squashing conceptually different logic - in this
case the libcst hooks needed to get code ranges versus the core business
logic of scraping annotations - into one place.

Unfortunately libcst kind of forces us to put the code range finding
functionality directly on the visitor because of the way it uses optional
metadata providers. But we can at least make this a mixin, because it is
conceptually very distinct from the annotation-finding code that uses it.

At this point I'm much happier with AnnotationCollector, which now only
includes code actually related to annotation collection - the stateful
context building has been pushed into a class with a semantic API and the
position logic is pushed into a mixin.

Reviewed By: grievejia

Differential Revision: D44377637

fbshipit-source-id: d6d33cd965b4b51242955bf3085aa23775d05a34